# Conflicts:
#	.idea/compiler.xml
#	.idea/modules.xml
#	Warlords.iml
#	src/main/java/com/ebicep/customentities/nms/pve/CustomEntitiesRegistry.java
#	src/main/java/com/ebicep/customentities/nms/pve/CustomEntity.java
#	src/main/java/com/ebicep/customentities/npc/NPCManager.java
#	src/main/java/com/ebicep/jda/BotManager.java
#	src/main/java/com/ebicep/warlords/Warlords.java
#	src/main/java/com/ebicep/warlords/abilties/ArcaneShieldPyromancer.java
#	src/main/java/com/ebicep/warlords/abilties/AvengersWrath.java
#	src/main/java/com/ebicep/warlords/abilties/Berserk.java
#	src/main/java/com/ebicep/warlords/abilties/BloodLust.java
#	src/main/java/com/ebicep/warlords/abilties/Boulder.java
#	src/main/java/com/ebicep/warlords/abilties/Consecrate.java
#	src/main/java/com/ebicep/warlords/abilties/CripplingStrike.java
#	src/main/java/com/ebicep/warlords/abilties/Earthliving.java
#	src/main/java/com/ebicep/warlords/abilties/HealingTotem.java
#	src/main/java/com/ebicep/warlords/abilties/Inferno.java
#	src/main/java/com/ebicep/warlords/abilties/InspiringPresence.java
#	src/main/java/com/ebicep/warlords/abilties/LightInfusionAvenger.java
#	src/main/java/com/ebicep/warlords/abilties/LightInfusionCrusader.java
#	src/main/java/com/ebicep/warlords/abilties/LightInfusionProtector.java
#	src/main/java/com/ebicep/warlords/abilties/OrbsOfLife.java
#	src/main/java/com/ebicep/warlords/abilties/OrderOfEviscerate.java
#	src/main/java/com/ebicep/warlords/abilties/PrismGuard.java
#	src/main/java/com/ebicep/warlords/abilties/Soulbinding.java
#	src/main/java/com/ebicep/warlords/abilties/TimeWarp.java
#	src/main/java/com/ebicep/warlords/abilties/TimeWarpAquamancer.java
#	src/main/java/com/ebicep/warlords/abilties/TimeWarpPyromancer.java
#	src/main/java/com/ebicep/warlords/abilties/UndyingArmy.java
#	src/main/java/com/ebicep/warlords/abilties/Windfury.java
#	src/main/java/com/ebicep/warlords/abilties/WoundingStrikeBerserker.java
#	src/main/java/com/ebicep/warlords/abilties/internal/AbstractStrikeBase.java
#	src/main/java/com/ebicep/warlords/abilties/pve/PveOrderOfEviscerate.java
#	src/main/java/com/ebicep/warlords/achievements/AchievementsMenu.java
#	src/main/java/com/ebicep/warlords/commands/CommandManager.java
#	src/main/java/com/ebicep/warlords/commands/debugcommands/game/GameStartCommand.java
#	src/main/java/com/ebicep/warlords/commands/debugcommands/ingame/DebugCommand.java
#	src/main/java/com/ebicep/warlords/commands/debugcommands/misc/OldTestCommand.java
#	src/main/java/com/ebicep/warlords/commands/debugcommands/misc/TestCommand.java
#	src/main/java/com/ebicep/warlords/commands/miscellaneouscommands/SpectateCommand.java
#	src/main/java/com/ebicep/warlords/database/leaderboards/stats/StatsLeaderboardManager.java
#	src/main/java/com/ebicep/warlords/database/leaderboards/stats/sections/AbstractStatsLeaderboardGameType.java
#	src/main/java/com/ebicep/warlords/database/repositories/events/pojos/GameEvents.java
#	src/main/java/com/ebicep/warlords/database/repositories/games/pojos/DatabaseGamePlayerBase.java
#	src/main/java/com/ebicep/warlords/database/repositories/games/pojos/pve/DatabaseGamePvEBase.java
#	src/main/java/com/ebicep/warlords/database/repositories/masterworksfair/pojos/MasterworksFair.java
#	src/main/java/com/ebicep/warlords/database/repositories/player/PlayersCollections.java
#	src/main/java/com/ebicep/warlords/database/repositories/player/pojos/general/DatabasePlayer.java
#	src/main/java/com/ebicep/warlords/database/repositories/player/pojos/general/DatabasePlayerCompStats.java
#	src/main/java/com/ebicep/warlords/database/repositories/player/pojos/general/DatabasePlayerPubStats.java
#	src/main/java/com/ebicep/warlords/database/repositories/player/pojos/general/TournamentStats.java
#	src/main/java/com/ebicep/warlords/database/repositories/player/pojos/pve/DatabasePlayerPvE.java
#	src/main/java/com/ebicep/warlords/database/repositories/player/pojos/pve/PvEDatabaseStatInformation.java
#	src/main/java/com/ebicep/warlords/events/WarlordsEvents.java
#	src/main/java/com/ebicep/warlords/events/player/ingame/WarlordsDamageHealingEvent.java
#	src/main/java/com/ebicep/warlords/events/player/ingame/pve/WarlordsGiveItemEvent.java
#	src/main/java/com/ebicep/warlords/game/GameMode.java
#	src/main/java/com/ebicep/warlords/game/option/AbstractScoreOnEventOption.java
#	src/main/java/com/ebicep/warlords/game/option/pve/onslaught/OnslaughtOption.java
#	src/main/java/com/ebicep/warlords/game/option/pve/tutorial/TutorialOption.java
#	src/main/java/com/ebicep/warlords/game/option/pve/wavedefense/WaveDefenseOption.java
#	src/main/java/com/ebicep/warlords/game/state/EndState.java
#	src/main/java/com/ebicep/warlords/guilds/commands/GuildDebugCommand.java
#	src/main/java/com/ebicep/warlords/menu/debugmenu/DebugMenuPlayerOptions.java
#	src/main/java/com/ebicep/warlords/party/commands/PartyCommand.java
#	src/main/java/com/ebicep/warlords/permissions/PermissionHandler.java
#	src/main/java/com/ebicep/warlords/permissions/Permissions.java
#	src/main/java/com/ebicep/warlords/player/general/CustomScoreboard.java
#	src/main/java/com/ebicep/warlords/player/ingame/WarlordsEntity.java
#	src/main/java/com/ebicep/warlords/player/ingame/WarlordsNPC.java
#	src/main/java/com/ebicep/warlords/pve/DifficultyMenu.java
#	src/main/java/com/ebicep/warlords/pve/commands/MobCommand.java
#	src/main/java/com/ebicep/warlords/pve/mobs/AbstractMob.java
#	src/main/java/com/ebicep/warlords/pve/mobs/blaze/Blaze.java
#	src/main/java/com/ebicep/warlords/pve/mobs/bosses/Boltaro.java
#	src/main/java/com/ebicep/warlords/pve/mobs/bosses/Chessking.java
#	src/main/java/com/ebicep/warlords/pve/mobs/bosses/Narmer.java
#	src/main/java/com/ebicep/warlords/pve/mobs/bosses/Torment.java
#	src/main/java/com/ebicep/warlords/pve/mobs/bosses/bossminions/BoltaroExiled.java
#	src/main/java/com/ebicep/warlords/pve/mobs/bosses/bossminions/BoltaroShadow.java
#	src/main/java/com/ebicep/warlords/pve/mobs/events/pharaohsrevenge/EventDjet.java
#	src/main/java/com/ebicep/warlords/pve/mobs/events/pharaohsrevenge/EventNarmer.java
#	src/main/java/com/ebicep/warlords/pve/mobs/pigzombie/VoidPigZombie.java
#	src/main/java/com/ebicep/warlords/pve/mobs/skeleton/ExiledSkeleton.java
#	src/main/java/com/ebicep/warlords/pve/mobs/slime/VoidSlime.java
#	src/main/java/com/ebicep/warlords/pve/mobs/witherskeleton/WitherWarrior.java
#	src/main/java/com/ebicep/warlords/pve/mobs/zombie/EliteZombie.java
#	src/main/java/com/ebicep/warlords/pve/mobs/zombie/EnvoyZombie.java
#	src/main/java/com/ebicep/warlords/pve/mobs/zombie/ExiledZombie.java
#	src/main/java/com/ebicep/warlords/pve/mobs/zombie/ForgottenZombie.java
#	src/main/java/com/ebicep/warlords/pve/mobs/zombie/VoidZombie.java
#	src/main/java/com/ebicep/warlords/pve/mobs/zombie/berserkzombie/EliteBerserkZombie.java
#	src/main/java/com/ebicep/warlords/pve/mobs/zombie/berserkzombie/EnvoyBerserkZombie.java
#	src/main/java/com/ebicep/warlords/pve/rewards/RewardInventory.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/AbstractUpgradeBranch.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/AutoUpgradeProfile.java
#	src/main/java/com/ebicep/warlords/pve/upgrades/paladin/avenger/AvengerStrikeBranch.java
#	src/main/java/com/ebicep/warlords/pve/weapons/AbstractWeapon.java
#	src/main/java/com/ebicep/warlords/pve/weapons/menu/WeaponManagerMenu.java
#	src/main/java/com/ebicep/warlords/pve/weapons/menu/WeaponTitleMenu.java
#	src/main/java/com/ebicep/warlords/pve/weapons/weapontypes/legendaries/LegendaryTitles.java
#	src/main/java/com/ebicep/warlords/util/pve/SkullUtils.java
#	src/main/resources/plugin.yml

+++ src/main/java/com/ebicep/warlords/player/ingame/PlayerStatisticsMinute.java
@@ -319,7 +319,7 @@ public class PlayerStatisticsMinute implements Iterable<PlayerStatisticsMinute.E
         }
 
         @Nonnegative
-        public float getRespawnTimeSpent() {
+        public int getRespawnTimeSpent() {
             return respawnTimeSpent;
         }
